+++ bin/.gitignore
@@ -1,0 +1,4 @@
+/Geometria_alakzat/
+/Main.class
+/Geometria_alakzatok/
+/counterHW/
