+++ src/Geometria_alakzatok/Parallelogram.java
@@ -1,4 +1,4 @@
-package Geometriai_alakzatok;
+package Geometria_alakzatok;
 
 public class Parallelogram extends Shape{
