+++ src/Geometria_alakzatok/Square.java
@@ -1,4 +1,4 @@
-package Geometriai_alakzatok;
+package Geometria_alakzatok;
 
 public class Square extends Shape{
