+++ src/Geometria_alakzatok/Triangle.java
@@ -1,4 +1,4 @@
-package Geometriai_alakzatok;
+package Geometria_alakzatok;
 
 public class Triangle extends Shape {
